+++ packages/databricks-vscode/src/whatsNewPopup.ts
@@ -36,12 +36,12 @@ export async function findFileFowWhatsNew(
 
 async function showV2Notification() {
     const choice = await window.showInformationMessage(
-        "Please note the V2 version of the extension will be made the default in a few days. With this update it is easier to set up your project, integrate with Databricks Asset Bundles and run your code remotely with fewer steps.",
+        "Version 2 of the extension will soon be made the default. With this update it is easier to set up your project, integrate with Databricks Asset Bundles, and run your code remotely.",
         "Learn more"
     );
     if (choice === "Learn more") {
         openExternal(
-            "https://docs.databricks.com/en/dev-tools/vscode-ext/install.html"
+            "https://docs.databricks.com/dev-tools/vscode-ext/index.html"
         );
     }
 }